feat(web-vitals): move to v2 folders
