+++ head.html
@@ -1,6 +1,6 @@
-<link rel="stylesheet" href="/style-v2.css"/>
-<script src="/scripts-v2.js" defer></script>
-<script src="/scripts/common-v2.js" type="module"></script>
+<link rel="stylesheet" href="/style/v2/style.css"/>
+<script src="/scripts/v2/scripts.js" defer></script>
+<script src="/scripts/v2/common.js" type="module"></script>
 
 <meta name="viewport" content="width=device-width, initial-scale=1"/>
