+++ scripts/v2/blank.js
@@ -12,7 +12,7 @@
 
 import {
   addClass,
-} from '/scripts/common-v2.js';
+} from '/scripts/v2/common.js';
 
 
 /**
