+++ scripts/v2/common.js
