+++ scripts/v2/post.js
@@ -17,7 +17,7 @@ import {
   createTag,
   extractTopicsAndProducts,
   globalPostLCP
-} from '/scripts/common-v2.js';
+} from '/scripts/v2/common.js';
 
 import {
   getTaxonomy
